Rename upgrade->update, self-upgrade->upgrade; require --all, bump to 0.4.0
